Input format corresponds to vertex value format

+++ giraph-examples/src/main/java/org/apache/giraph/examples/DeltaVertexWritable.java
@@ -55,7 +55,7 @@ class DeltaVertexWritable implements Writable {
     
     public double getDist() { return this.dist; }
     
-    public double getBucket() { return this.bucket; }
+    public long getBucket() { return this.bucket; }
     
     public boolean isProcessing() { return this.doneLight==0?true:false; }
